finish task2, semaphore, barrier and mailbox

+++ UCAS_OS/Project3-Interactive_OS_and_Process_Management/test/test_project3/test_barrier.c
@@ -16,12 +16,10 @@ void test_barrier(void)
 
     mthread_barrier_init(&barrier, NUM_TB);
 
-    struct task_info child_task = {(uintptr_t)&barrier_task,
-                                   USER_PROCESS};
+    struct task_info child_task = {(uintptr_t)&barrier_task, USER_PROCESS};
     pid_t pids[NUM_TB];
     for (int i = 0; i < NUM_TB; ++i) {
-        pids[i] = sys_spawn(&child_task, (void*)(i + PRINT_START),
-                            ENTER_ZOMBIE_ON_EXIT);
+        pids[i] = sys_spawn(&child_task, (void*)(long)(i + PRINT_START), ENTER_ZOMBIE_ON_EXIT);
     }
 
     for (int i = 0; i < NUM_TB; ++i) {
@@ -50,6 +48,5 @@ void barrier_task(int print_location)
                i, next);
         sys_sleep(next);
     }
-
     sys_exit();
 }
